# docs/openapi.yaml
openapi: 3.0.3
info:
  title: E-Bookstore API
  version: 1.0.0
  description: REST API for the E-Bookstore e-commerce platform
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
...
paths:
...
  /users/me:
    get:
      tags: [Users]
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    patch:
      tags: [Users]
      summary: Update current user profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

  # ── ADDRESSES ─────────────────────────────────────────────────────────────
  /users/me/addresses:
    get:
      tags: [Addresses]
      summary: List delivery addresses
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Address list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Address'
    post:
      tags: [Addresses]
      summary: Add a delivery address
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddressRequest'
      responses:
        '201':
          description: Address created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Address'
  /users/me/addresses/{addressId}:
    patch:
      tags: [Addresses]
      summary: Update an address
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/addressId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddressRequest'
      responses:
        '200':
          description: Updated address
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Address'
    delete:
      tags: [Addresses]
      summary: Delete an address
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/addressId'
      responses:
        '204':
          description: Deleted

  # ── CATEGORIES ────────────────────────────────────────────────────────────
...
  /payments:
    post:
      tags: [Payments]
      summary: Process payment for an order
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentRequest'
      responses:
        '201':
          description: Payment processed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
        '402':
          $ref: '#/components/responses/PaymentRequired'
  /payments/{paymentId}:
    get:
      tags: [Payments]
      summary: Get payment status
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/paymentId'
      responses:
        '200':
          description: Payment detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'

# ── COMPONENTS ────────────────────────────────────────────────────────────────
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    addressId:
      name: addressId
      in: path
      required: true
      schema: { type: string, format: uuid }
    paymentId:
      name: paymentId
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
...
    PaymentRequired:
      description: Payment failed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
  schemas:
    ApiError:
      type: object
      properties:
        code: { type: string }
        message: { type: string }
...
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string }
        firstName: { type: string }
        lastName: { type: string }
        giftPoints: { type: integer }
    UpdateProfileRequest:
      type: object
      properties:
        firstName: { type: string }
        lastName: { type: string }
    Address:
      type: object
      properties:
        id: { type: string, format: uuid }
        label: { type: string }
        line1: { type: string }
        line2: { type: string }
        city: { type: string }
        state: { type: string }
        postalCode: { type: string }
        country: { type: string }
        isDefault: { type: boolean }
    AddressRequest:
      type: object
      required: [label, line1, city, state, postalCode, country]
      properties:
        label: { type: string }
        line1: { type: string }
        line2: { type: string }
        city: { type: string }
        state: { type: string }
        postalCode: { type: string }
        country: { type: string }
        isDefault: { type: boolean }
...
    PaymentRequest:
      type: object
      required: [orderId, method]
      properties:
        orderId: { type: string, format: uuid }
        method:
          type: string
          enum: [card, paypal, gift_points]
        cardToken:
          type: string
          description: Tokenised card reference (card method only)
    Payment:
      type: object
      properties:
        id: { type: string, format: uuid }
        orderId: { type: string, format: uuid }
        method: { type: string }
        status:
          type: string
          enum: [pending, completed, failed, refunded]
        confirmationRef: { type: string }
        amount: { type: number }
        paidAt: { type: string, format: date-time }